+++ packages/wise_feedback/lib/wise_feedback.dart
@@ -4,6 +4,7 @@ library wise_feedback;
 export 'src/metadata/metadata.dart';
 export 'src/models/models.dart';
 export 'src/screens/screens.dart';
+export 'src/template/template.dart';
 export 'src/theme/theme.dart';
 export 'src/transport/transport.dart';
 export 'src/widgets/widgets.dart';